+++ cmd/main/cmd_publish.go
@@ -3,10 +3,10 @@
 package main
 
 import (
-	"bytes"
 	"crypto/tls"
 	"encoding/json"
 	"io"
+	"io/ioutil"
 
 	"github.com/jandelgado/rabtap/pkg"
 	"github.com/streadway/amqp"
@@ -21,9 +21,10 @@ type CmdPublishArg struct {
 	readNextMessageFunc MessageReaderFunc
 }
 
-// MessageReaderFunc provides messages that can be sent to an exchange. If no
-// more messages are available io.EOF is returned as error.
-type MessageReaderFunc func() (amqp.Publishing, error)
+// MessageReaderFunc provides messages that can be sent to an exchange.
+// returns the message to be published, a flag if more messages are to be read,
+// and an error.
+type MessageReaderFunc func() (amqp.Publishing, bool, error)
 
 // SignalChannel transports os.Signal objects like e.g. os.Interrupt
 // type SignalChannel chan os.Signal
@@ -46,38 +47,32 @@ func publishMessage(publishChannel rabtap.PublishChannel,
 // readSingleMessageFromRawFile reads a single messages from the given io.Reader
 // which is typically stdin or a file. If reading from stdin, CTRL+D (linux)
 // or CTRL+Z (Win) on an empty line terminates the reader.
-func readSingleMessageFromRawFile(reader io.Reader) (amqp.Publishing, error) {
-	buf := new(bytes.Buffer)
-	numRead, err := buf.ReadFrom(reader)
-	if err != nil {
-		return amqp.Publishing{}, err
-	} else if numRead == 0 {
-		return amqp.Publishing{}, io.EOF
-	}
-	return amqp.Publishing{Body: buf.Bytes()}, nil
+func readSingleMessageFromRawFile(reader io.Reader) (amqp.Publishing, bool, error) {
+	buf, err := ioutil.ReadAll(reader)
+	return amqp.Publishing{Body: buf}, false, err
 }
 
 // readNextMessageFromJSONStream reads JSON messages from the given decoder as long
 // as there are messages available.
-func readNextMessageFromJSONStream(decoder *json.Decoder) (amqp.Publishing, error) {
+func readNextMessageFromJSONStream(decoder *json.Decoder) (amqp.Publishing, bool, error) {
 	message := RabtapPersistentMessage{}
 	err := decoder.Decode(&message)
 	if err != nil {
-		return amqp.Publishing{}, err
+		return amqp.Publishing{}, false, err
 	}
-	return message.ToAmqpPublishing(), nil
+	return message.ToAmqpPublishing(), true, nil
 }
 
 // createMessageReaderFunc returns a function that reads messages from the
-// the given reader in JSON or raw-format
+// the given reader in JSON or raw-format TODO drop boolean param
 func createMessageReaderFunc(jsonFormat bool, reader io.Reader) MessageReaderFunc {
 	if jsonFormat {
 		decoder := json.NewDecoder(reader)
-		return func() (amqp.Publishing, error) {
+		return func() (amqp.Publishing, bool, error) {
 			return readNextMessageFromJSONStream(decoder)
 		}
 	}
-	return func() (amqp.Publishing, error) {
+	return func() (amqp.Publishing, bool, error) {
 		return readSingleMessageFromRawFile(reader)
 	}
 }
@@ -87,15 +82,23 @@ func createMessageReaderFunc(jsonFormat bool, reader io.Reader) MessageReaderFun
 func publishMessageStream(publishChannel rabtap.PublishChannel,
 	exchange, routingKey string,
 	readNextMessageFunc MessageReaderFunc) error {
+
 	for {
-		amqpMessage, err := readNextMessageFunc()
+		msg, more, err := readNextMessageFunc()
 		switch err {
 		case io.EOF:
+			close(publishChannel)
 			return nil
 		case nil:
-			publishMessage(publishChannel, exchange, routingKey, amqpMessage)
+			publishMessage(publishChannel, exchange, routingKey, msg)
 		default:
+			close(publishChannel)
 			return err
+		}
+
+		if !more {
+			close(publishChannel)
+			return nil
 		}
 	}
 }
@@ -108,7 +111,6 @@ func cmdPublish(cmd CmdPublishArg) error {
 	publisher := rabtap.NewAmqpPublish(cmd.amqpURI, cmd.tlsConfig, log)
 	defer publisher.Close()
 	publishChannel := make(rabtap.PublishChannel)
-	go publisher.EstablishConnection(publishChannel)
-	return publishMessageStream(publishChannel, cmd.exchange, cmd.routingKey,
-		cmd.readNextMessageFunc)
+	go publishMessageStream(publishChannel, cmd.exchange, cmd.routingKey, cmd.readNextMessageFunc)
+	return publisher.EstablishConnection(publishChannel)
 }